+++ kmjxc/dba/Back_Stock.cs
@@ -14,12 +14,12 @@ namespace KM.JXC.DBA
     
     public partial class Back_Stock
     {
-        public long Back_Sock_ID { get; set; }
+        public int Back_Sock_ID { get; set; }
         public long Back_Date { get; set; }
-        public long Shop_ID { get; set; }
-        public long StoreHouse_ID { get; set; }
-        public long User_ID { get; set; }
+        public int Shop_ID { get; set; }
+        public int StoreHouse_ID { get; set; }
+        public int User_ID { get; set; }
         public string Description { get; set; }
-        public long Back_Sale_ID { get; set; }
+        public int Back_Sale_ID { get; set; }
     }
 }
