+++ kmjxc/dba/Buy_Detail.cs
@@ -14,10 +14,10 @@ namespace KM.JXC.DBA
     
     public partial class Buy_Detail
     {
-        public long Buy_ID { get; set; }
-        public long Product_ID { get; set; }
-        public long Buy_Order_ID { get; set; }
-        public long Quantity { get; set; }
+        public int Buy_ID { get; set; }
+        public int Product_ID { get; set; }
+        public int Buy_Order_ID { get; set; }
+        public int Quantity { get; set; }
         public decimal Price { get; set; }
         public long Create_Date { get; set; }
     }
