+++ kmjxc/dba/Employee.cs
@@ -14,7 +14,7 @@ namespace KM.JXC.DBA
     
     public partial class Employee
     {
-        public long Employee_ID { get; set; }
+        public int Employee_ID { get; set; }
         public string Department { get; set; }
         public string Name { get; set; }
         public string Duty { get; set; }
@@ -26,6 +26,6 @@ namespace KM.JXC.DBA
         public string Address { get; set; }
         public string Phone { get; set; }
         public string Email { get; set; }
-        public Nullable<long> User_ID { get; set; }
+        public Nullable<int> User_ID { get; set; }
     }
 }
