+++ kmjxc/dba/Product_Class.cs
@@ -14,13 +14,13 @@ namespace KM.JXC.DBA
     
     public partial class Product_Class
     {
-        public long Product_Class_ID { get; set; }
+        public int Product_Class_ID { get; set; }
         public string Name { get; set; }
-        public Nullable<long> Parent_ID { get; set; }
+        public Nullable<int> Parent_ID { get; set; }
         public Nullable<byte> Order { get; set; }
         public long Create_Time { get; set; }
-        public long Create_User_ID { get; set; }
-        public long Shop_ID { get; set; }
+        public int Create_User_ID { get; set; }
+        public int Shop_ID { get; set; }
         public bool Enabled { get; set; }
     }
 }
