+++ kmjxc/dba/Sale_Detail.cs
@@ -14,14 +14,14 @@ namespace KM.JXC.DBA
     
     public partial class Sale_Detail
     {
-        public long Sale_ID { get; set; }
-        public long Product_ID { get; set; }
+        public int Sale_ID { get; set; }
+        public int Product_ID { get; set; }
         public string Mall_Order_ID { get; set; }
-        public long Quantity { get; set; }
+        public int Quantity { get; set; }
         public decimal Price { get; set; }
         public decimal Discount { get; set; }
         public bool Status { get; set; }
-        public Nullable<int> Amount { get; set; }
+        public Nullable<decimal> Amount { get; set; }
         public int Supplier_ID { get; set; }
     }
 }
