+++ kmjxc/dba/Shop.cs
@@ -14,12 +14,12 @@ namespace KM.JXC.DBA
     
     public partial class Shop
     {
-        public long Shop_ID { get; set; }
+        public int Shop_ID { get; set; }
         public string Mall_Shop_ID { get; set; }
         public string Name { get; set; }
-        public long User_ID { get; set; }
-        public Nullable<long> Parent_Shop_ID { get; set; }
+        public int User_ID { get; set; }
+        public Nullable<int> Parent_Shop_ID { get; set; }
         public string Description { get; set; }
-        public long Mall_Type_ID { get; set; }
+        public int Mall_Type_ID { get; set; }
     }
 }
